Add interpolation method as command line argument
Current options are numpy and scipy. Scipy is what we've been doing so
far: a cubic interpolation using scipy's interp1d module. Now, we have
the option to use numpy's interp which is a 1D linear interpolation. The
linear interpolation is (presumably) less accurate but significantly
faster, especially since we can compile the interpolation function with
numba.